Restructured Public Courses files also did minor CSS improvements
