+++ app-next/app/public-courses/page.js
@@ -3,3 +3,11 @@ import PublicCourses from "@/components/PublicCourses/page";
 export default function NestedPage() {
   return <PublicCourses/>;
 }
+
+
+
+// export default function PublicCourses() {
+//   return <div>
+//     {/* - - - - - - -Border - - - - - - - */}
+//   </div>;
+// }
