+++ app-next/components/Footer.js/Footer.module.css
@@ -38,6 +38,7 @@
   justify-content: center;
 }
 .sectionTitle {
+  padding-right: 6rem;
   font-size: 1.125rem;
   font-weight: 600;
   color: #111827;
